+++ packages/explorer/src/app/(explorer)/[chainName]/worlds/[worldAddress]/interact/Form.tsx
@@ -6,14 +6,14 @@ import { useDeferredValue, useState } from "react";
 import { Input } from "../../../../../../components/ui/Input";
 import { Separator } from "../../../../../../components/ui/Separator";
 import { Skeleton } from "../../../../../../components/ui/Skeleton";
-import { useHashState } from "../../../../../../hooks/useHashState";
-import { cn } from "../../../../../../lib/utils";
-import { useAbiQuery } from "../../../../../../queries/useAbiQuery";
+import { cn } from "../../../../../../utils";
+import { useHashState } from "../../../../hooks/useHashState";
+import { useWorldAbiQuery } from "../../../../queries/useWorldAbiQuery";
 import { FunctionField } from "./FunctionField";
 
 export function Form() {
   const [hash] = useHashState();
-  const { data, isFetched } = useAbiQuery();
+  const { data, isFetched } = useWorldAbiQuery();
   const [filterValue, setFilterValue] = useState("");
   const deferredFilterValue = useDeferredValue(filterValue);
   const filteredFunctions = data?.abi?.filter((item) =>